Removed unused methods Moved more common logic from individual controllers to their controller base
