+++ src/TheReplacement.PTA.Services.Core/Extensions/RequestExtensions.cs
@@ -120,7 +120,7 @@ namespace TheReplacement.PTA.Services.Core.Extensions
                 pokemonModel.OriginalTrainerId = trainer.TrainerId;
                 pokemonModel.TrainerId = trainer.TrainerId;
                 pokemonModel.GameId = trainer.GameId;
-                pokemonModel.Pokeball = Pokeball.Basic_Ball.ToString();
+                pokemonModel.Pokeball = Pokeball.Basic_Ball.ToString().Replace("_", "");
                 DatabaseUtility.TryAddPokemon(pokemonModel, out _);
                 var game = DatabaseUtility.FindGame(trainer.GameId);
                 var caughtPokemonLog = new LogModel
@@ -138,52 +138,6 @@ namespace TheReplacement.PTA.Services.Core.Extensions
                     DatabaseUtility.UpdateDexItemIsCaught(trainer.TrainerId, pokemonModel.DexNo);
                 }
             }
-        }
-
-        public static (PokemonModel Pokemon, AbstractMessage Message) BuildPokemon(
-            this HttpRequest request,
-            Guid trainerId,
-            Guid gameId,
-            WildPokemon wild)
-        {
-            var (pokemon, error) = BuildDefaultPokemon(wild);
-            if (pokemon == null)
-            {
-                return (null, error);
-            }
-
-            pokemon.TrainerId = trainerId;
-            pokemon.OriginalTrainerId = trainerId;
-            pokemon.GameId = gameId;
-            return (pokemon, null);
-        }
-
-        public static async Task<(string GamePassword, string GameMasterUsername, string GamemasterPassword, IEnumerable<string> Errors)> GetStartGameCredentials(
-            this HttpRequest request)
-        {
-            var body = await request.GetRequestBody();
-            var errors = new[] { "gameSessionPassword", "gmUsername", "gmPassword" }
-                .Select(key => body[key] != null ? null : $"Missing {key}")
-                .Where(error => error != null);
-            var gamePassword = (string)body["gameSessionPassword"];
-            var username = (string)body["gmUsername"];
-            var password = (string)body["gmPassword"];
-
-            return (gamePassword, username, password, errors);
-        }
-
-        public static async Task<(string GameId, string Username, string Password, IEnumerable<string> Errors)> GetTrainerCredentials(
-            this HttpRequest request)
-        {
-            var body = await request.GetRequestBody();
-            var errors = new[] { "trainerName", "password" }
-                .Select(key => body[key] != null ? null : $"Missing {key}")
-                .Where(error => error != null);
-            var gameId = request.Query["gameId"];
-            var trainerName = (string)body["trainerName"];
-            var password = (string)body["password"];
-
-            return (gameId, trainerName, password, errors);
         }
 
         public static async Task<(string Username, string Password, IEnumerable<string> Errors)> GetUserCredentials(
@@ -232,67 +186,6 @@ namespace TheReplacement.PTA.Services.Core.Extensions
             };
 
             return Array.Empty<Guid>();
-        }
-
-        private static async Task<(bool IsValid, InvalidQueryStringMessage Message)>  ValidateMandatoryPokemonKeys(
-            HttpRequest request)
-        {
-            var body = await request.GetRequestBody();
-            var fails = MandatoryPokemonKeys.Where(key => string.IsNullOrWhiteSpace((string)body[key]));
-            var error = new InvalidQueryStringMessage
-            {
-                MissingParameters = fails
-            };
-            return (fails.Any(), error);
-        }
-
-        private static (PokemonModel Pokemon, AbstractMessage Message) BuildDefaultPokemon(WildPokemon wild)
-        {
-            var random = new Random();
-            if (!Enum.TryParse(wild.Gender, true, out Gender gender))
-            {
-                var genders = Enum.GetValues<Gender>();
-                gender = genders[random.Next(genders.Length)];
-            }
-
-            if (!Enum.TryParse(wild.Nature, true, out Nature nature))
-            {
-                var natures = Enum.GetValues<Nature>();
-                nature = natures[random.Next(1, natures.Length)];
-            }
-
-            if (!Enum.TryParse(wild.Status, true, out Status status))
-            {
-                status = Status.Normal;
-            }
-
-            var form = wild.Form.Replace('_', '/');
-            if (string.IsNullOrWhiteSpace(form))
-            {
-                return (null, new GenericMessage($"Mission form in request"));
-            }
-
-            var pokemon = DexUtility.GetNewPokemon
-            (
-                wild.Pokemon,
-                nature,
-                gender,
-                status,
-                null,
-                form
-            );
-
-            if (pokemon == null)
-            {
-                return (null, new GenericMessage("Failed to build pokemon"));
-            }
-
-            if (wild.ForceShiny)
-            {
-                pokemon.IsShiny = true;
-            }
-            pokemon.Pokeball = Pokeball.Basic_Ball.ToString().Replace("_", "");
-            return (pokemon, null);
         }
     }
 }
